dom/contacts/tests/test_contacts_basics.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id={674720}
-->
<head>
  <title>Test for Bug {674720} WebContacts</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id={674720}">Mozilla Bug {674720}</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

"use strict"

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
Components.classes["@mozilla.org/permissionmanager;1"]
          .getService(Components.interfaces.nsIPermissionManager)
          .add(SpecialPowers.getDocumentURIObject(window.document),
               "webcontacts-manage",
               Components.interfaces.nsIPermissionManager.ALLOW_ACTION);

var adr1 = {
  streetAddress: "street 1",
  locality: "locality 1",
  region: "region 1",
  postalCode: "postal code 1",
  countryName: "country 1"
};

var adr2 = {
  streetAddress: "street2",
  locality: "locality2",
  region: "region2",
  postalCode: "postal code2",
  countryName: "country2"
};

var properties1 = {
  name: "Testname1",
  familyName: "TestFamilyName",
  givenName: ["Test1","Test2"],
  nickname: "nicktest",
  tel: ["123456"],
  adr: adr1
};

var properties2 = {
  name: ["dummyName", "dummyName2"],
  familyName: "dummyFamilyName",
  givenName: "dummyGivenName",
  honorificPrefix: ["dummyHonorificPrefix","dummyHonorificPrefix2"],
  honorificSuffix: "dummyHonorificSuffix",
  additionalName: "dummyadditionalName",
  nickname: "dummyNickname",
  tel: ["123456789", "234567890"],
  email: ["a@b.c", "b@c.d"],
  adr: [adr1, adr2],
  impp: ["im1", "im2"],
  org: ["org1", "org2"],
  bday: new Date("1980, 12, 01"),
  note: "test note",
  photo: ["pic1", "pic2"],
  category: ["cat1", "cat2"],
  url: ["www.1.com", "www.2.com"],
  anniversary: new Date("2000, 12, 01"),
  sex: "male",
  genderIdentity: "test"
};

var sample_id1;
var sample_id2;

var createResult1;
var createResult2;

var findResult1;
var findResult2;

function clearTemps() {
  sample_id1 = null;
  sample_id2 = null;
  createResult1 = null;
  createResult2 = null;
  findResult1 = null;
  findResult2 = null;
}

function onUnwantedSuccess() {
  ok(false, "onUnwantedSuccess: shouldn't get here");
}

function onFailure() {
  ok(false, "in on Failure!");
}

function checkStr(str1, str2, msg) {
  if (str1)
    ok(typeof str1 == "string" ? [str1] : str1, (typeof str2 == "string") ? [str2] : str2, msg);
}

function checkAddress(adr1, adr2) {
  checkStr(adr1.streetAddress, adr2.streetAddress, "Same streetAddress");
  checkStr(adr1.locality, adr2.locality, "Same locality");
  checkStr(adr1.region, adr2.region, "Same region");
  checkStr(adr1.postalCode, adr2.postalCode, "Same postalCode");
  checkStr(adr1.countryName, adr2.countryName, "Same countryName");
}

function checkContacts(contact1, contact2) {
  checkStr(contact1.name, contact2.name, "Same name");
  checkStr(contact1.honorificPrefix, contact2.honorificPrefix, "Same honorificPrefix");
  checkStr(contact1.givenName, contact2.givenName, "Same givenName");
  checkStr(contact1.additionalName, contact2.additionalName, "Same additionalName");
  checkStr(contact1.familyName, contact2.familyName, "Same familyName");
  checkStr(contact1.honorificSuffix, contact2.honorificSuffix, "Same honorificSuffix");
  checkStr(contact1.nickname, contact2.nickname, "Same nickname");
  checkStr(contact1.email, contact2.email, "Same email");
  checkStr(contact1.photo, contact2.photo, "Same photo");
  checkStr(contact1.url, contact2.url, "Same url");
  checkStr(contact1.category, contact2.category, "Same category");
  is(contact1.bday ? contact1.bday.valueOf() : null, contact2.bday ? contact2.bday.valueOf() : null, "Same birthday");
  checkStr(contact1.note, contact2.note, "Same note");
  checkStr(contact1.impp, contact2.impp, "Same impp");
  is(contact1.anniversary ? contact1.anniversary.valueOf() : null , contact2.anniversary ? contact2.anniversary.valueOf() : null, "Same anniversary");
  is(contact1.sex, contact2.sex, "Same sex");
  is(contact1.genderIdentity, contact2.genderIdentity, "Same genderIdentity");

  for (var i in contact1.adr)
    checkAddress(contact1.adr[i], contact2.adr[i]);
}

var req;
var index = 0;

var mozContacts = window.navigator.mozContacts

var steps = [
  function () {
    ok(true, "Deleting database");
    req = mozContacts.clear();
    req.onsuccess = function () {
      ok(true, "Deleted the database");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find({});
    req.onsuccess = function () {
      ok(req.result.length == 0, "Empty database is empty.");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding empty contact");
    createResult1 = new mozContact();
    createResult1.init({});
    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find({});
    req.onsuccess = function () {
      ok(req.result.length == 1, "One contact.");
      findResult1 = req.result[0];
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting empty contact");
    req = navigator.mozContacts.remove(findResult1);
    req.onsuccess = function () {
      var req2 = mozContacts.find({});
      req2.onsuccess = function () {
        ok(req2.result.length == 0, "Empty Database.");
        clearTemps();
        next();
      }
      req2.onerror = onFailure;
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding a new contact1");
    createResult1 = new mozContact();
    createResult1.init(properties1);
    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      checkContacts(properties1, createResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find({});
    req.onsuccess = function() {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, findResult1);
      ok(findResult1.updated, "Has updated field");
      ok(findResult1.published, "Has published field");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Modifying contact1");
    findResult1.impp = properties1.impp = (["phil impp"]);
    req = navigator.mozContacts.save(findResult1);
    req.onsuccess = function () {
      var req2 = mozContacts.find({});
      req2.onsuccess = function() {
        ok(req2.result.length == 1, "Found exactly 1 contact.");
        findResult2 = req2.result[0];
        ok(findResult2.id == sample_id1, "Same ID");
        checkContacts(findResult2, properties1);
        ok(findResult2.impp.length == 1, "Found exactly 1 IMS info.");
        next();
      };
      req2.onerror = onFailure;
    };
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Saving old contact, should abort!");
    req = mozContacts.save(createResult1);
    req.onsuccess = onUnwantedSuccess;
    req.onerror   = function() { ok(true, "Successfully declined updating old contact!"); next(); };
  },
  function() {
    ok(true, "Saving old contact, should abort!");
    req = mozContacts.save(findResult1)
    req.onsuccess = onUnwantedSuccess;
    req.onerror   = function() { ok(true, "Successfully declined updating old contact!"); next(); };
  },
  function () {
    ok(true, "Retrieving a specific contact by ID");
    var options = {filterBy: ["id"],
                   filterOp: "equals",
                   filterValue: sample_id1};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving a specific contact by givenName");
    var options = {filterBy: ["givenName"],
                   filterOp: "equals",
                   filterValue: properties1.givenName[0]};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, createResult1);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Modifying contact2");
    findResult1.impp = properties1.impp = (["phil impp"]);
    req = mozContacts.save(findResult1);
    req.onsuccess = function () {
      var req2 = mozContacts.find({});
      req2.onsuccess = function () {
        ok(req2.result.length == 1, "Found exactly 1 contact.");
        findResult1 = req2.result[0];
        ok(findResult1.id == sample_id1, "Same ID");
        checkContacts(findResult1, createResult1);
        ok(findResult1.impp.length == 1, "Found exactly 1 IMS info.");
        next();
      }
      req2.onerror = onFailure;
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts by query");
    var options = {filterBy: ["name", "email"],
                   filterOp: "icontains",
                   filterValue: properties1.name[0].substring(0,4)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, createResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts by query");
    var options = {filterBy: ["nickname", "email"],
                   filterOp: "icontains",
                   filterValue: properties1.nickname};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, createResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts with multiple indices");
    var options = {filterBy: ["nickname", "email", "name"],
                   filterOp: "equals",
                   filterValue: properties1.nickname};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, createResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Modifying contact3");
    findResult1.email = (properties1.nickname);
    findResult1.nickname = "TEST";
    var newContact = new mozContact();
    newContact.init(findResult1);
    req = mozContacts.save(newContact);
    req.onsuccess = function () {
      var options = {filterBy: ["nickname", "email", "name"],
                     filterOp: "equals",
                     filterValue: properties1.nickname};
      // One contact has it in nickname and the other in email
      var req2 = mozContacts.find(options);
      req2.onsuccess = function () {
        ok(req2.result.length == 2, "Found exactly 2 contacts.");
        ok(req2.result[0].id != req2.result[1].id, "Different ID");
        next();
      }
      req2.onerror = onFailure;
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting contact" + findResult1.id);
    req = mozContacts.remove(findResult1);
    req.onsuccess = function () {
      var req2 = mozContacts.find({});
      req2.onsuccess = function () {
        ok(req2.result.length == 1, "One contact left.");
        findResult1 = req2.result[0];
        next();
      }
      req2.onerror = onFailure;
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    req = mozContacts.remove(findResult1);
    req.onsuccess =  function () {
      clearTemps();
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding a new contact");
    createResult1 = new mozContact();
    createResult1.init(properties1);
    req = mozContacts.save(createResult1)
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding a new contact2");
    createResult2 = new mozContact();
    createResult2.init(properties2);
    req = mozContacts.save(createResult2);
    req.onsuccess = function () {
      ok(createResult2.id, "The contact now has an ID.");
      sample_id2 = createResult2.id;
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find({})
    req.onsuccess = function () {
      ok(req.result.length == 2, "Found exactly 2 contact.");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    console.log("Searching contacts by query1");
    var options = {filterBy: ["name", "email"],
                   filterOp: "icontains",
                   filterValue: properties1.name[0].substring(0, 4)}
    req = mozContacts.find(options)
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, createResult1);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts by query2");
    var options = {filterBy: ["name", "email"],
                   filterOp: "icontains",
                   filterValue: properties2.name[0].substring(0, 4)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.adr.length == 2, "Adr length 2");
      checkContacts(findResult1, createResult2);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts by tel");
    var options = {filterBy: ["tel"],
                   filterOp: "contains",
                   filterValue: properties2.tel[0].substring(0, 7)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id2, "Same ID");
      checkContacts(findResult1, createResult2);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts by email");
    var options = {filterBy: ["email"],
                   filterOp: "contains",
                   filterValue: properties2.email[0].substring(0, 4)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id2, "Same ID");
      checkContacts(findResult1, createResult2);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    req = mozContacts.clear();
    req.onsuccess = function () {
      clearTemps();
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding 100 contacts");
    for (var i=0; i<99; i++) {
      createResult1 = new mozContact();
      createResult1.init(properties1);
      req = mozContacts.save(createResult1);
      req.onsuccess = function () {
        ok(createResult1.id, "The contact now has an ID.");
      };
      req.onerror = onFailure;
    };
    createResult1 = new mozContact();
    createResult1.init(properties1);
    req = mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      ok(createResult1.name == properties1.name, "Same Name");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find({});
    req.onsuccess = function () {
      ok(req.result.length == 100, "100 Entries.");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts with limit 10");
    var options = { filterLimit: 10 };
    req = mozContacts.find(options);
    req.onsuccess = function () {
      ok(req.result.length == 10, "10 Entries.");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts2");
    var options = {filterBy: ["name"],
                   filterOp: "icontains",
                   filterValue: properties2.name[0].substring(0, 4)};
    req = mozContacts.find({});
    req.onsuccess = function () {
      ok(req.result.length == 100, "100 Entries.");
      checkContacts(createResult1, req.result[99]);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    req = mozContacts.clear();
    req.onsuccess = function () {
      clearTemps();
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Testing clone contact");
    createResult1 = new mozContact();
    createResult1.init(properties1);
    req = mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      ok(createResult1.name == properties1.name, "Same Name");
      next();
    }
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Testing clone contact2");
    var cloned = new mozContact(createResult1);
    ok(cloned.id != createResult1.id, "Cloned contact has new ID");
    cloned.email = "new email!";
    cloned.givenName = "Tom";
    req = mozContacts.save(cloned);
    req.onsuccess = function () {
      ok(cloned.id, "The contact now has an ID.");
      ok(cloned.email == "new email!", "Same Email");
      ok(createResult1.email != cloned.email, "Clone has different email");
      ok(cloned.givenName == "Tom", "New Name");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    var options = {filterBy: ["name"],
                   filterOp: "icontains",
                   filterValue: properties2.name[0].substring(0, 4)};
    req = mozContacts.find({});
    req.onsuccess = function () {
      ok(req.result.length == 2, "2 Entries.");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Search with redundant fields should only return 1 contact");
    createResult1 = new mozContact();
    createResult1.init({name: "XXX", nickname: "XXX", email: "XXX", tel: "XXX"});
    req = mozContacts.save(createResult1);
    req.onsuccess = function() {
      var options = {filterBy: [],
                     filterOp: "equals",
                     filterValue: "XXX"};
      var req2 = mozContacts.find(options);
      req2.onsuccess = function() {
        ok(req2.result.length == 1, "1 Entry");
        next();
      }
      req2.onerror = onFailure;
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    req = mozContacts.clear()
    req.onsuccess = function () {
      ok(true, "Deleted the database");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "all done!\n");
    clearTemps();

    SimpleTest.finish();
  }
];

function next() {
  ok(true, "Begin!");
  if (index >= steps.length) {
    ok(false, "Shouldn't get here!");
    return;
  }
  try {
    steps[index]();
  } catch(ex) {
    ok(false, "Caught exception", ex);
  }
  index += 1;
}

function permissionTest() {
  if (gContactsEnabled) {
    next();
  } else {
    is(mozContacts, null, "mozContacts is null when not enabled.");
    SimpleTest.finish();
  }
}

var gContactsEnabled = SpecialPowers.getBoolPref("dom.mozContacts.enabled");
SimpleTest.waitForExplicitFinish();
addLoadEvent(permissionTest);

ok(true, "test passed");
</script>
</pre>
</body>
</html>
